revert to i16s

+++ packages/input/src/events.ts
@@ -2,8 +2,6 @@ import { struct } from "thyseus";
 
 import { Key, PointerType } from "./types";
 
-// TODO: change i32s back to i16
-
 @struct
 export class MouseEvent {
   @struct.bool declare altKey: boolean;
@@ -13,8 +11,8 @@ export class MouseEvent {
   @struct.u16 declare clientY: number;
   @struct.bool declare ctrlKey: boolean;
   @struct.bool declare metaKey: boolean;
-  @struct.i32 declare movementX: number;
-  @struct.i32 declare movementY: number;
+  @struct.i16 declare movementX: number;
+  @struct.i16 declare movementY: number;
   @struct.u16 declare screenX: number;
   @struct.u16 declare screenY: number;
   @struct.bool declare shiftKey: boolean;
@@ -32,8 +30,8 @@ export class PointerEvent {
   @struct.u16 declare clientY: number;
   @struct.bool declare ctrlKey: boolean;
   @struct.bool declare metaKey: boolean;
-  @struct.i32 declare movementX: number;
-  @struct.i32 declare movementY: number;
+  @struct.i16 declare movementX: number;
+  @struct.i16 declare movementY: number;
   @struct.u16 declare screenX: number;
   @struct.u16 declare screenY: number;
   @struct.bool declare shiftKey: boolean;
